refactor(devtools): split the MCP workbench router from its server
The CLI workbench already separates its router from the HTTP server it
runs on. The MCP workbench now matches: `createAttachedRouter` owns the
browser sessions, the JSON API, the OAuth callback, and the shell, and
`startAttachedDevtoolsServer` only binds loopback and closes cleanly.

The router takes its API mount point as an option so the same handler can
be mounted next to another workbench on one origin.
